WIP: separating out automation.py into submit and copy only
