--- data-prep/Sentinel/automate/config.ini
@@ -1,15 +1,0 @@
-[HyP3]
-username = 
-password = 
-[aws]
-dest_bucket = servir-public
-prefix_str = geotiffs/peru/sentinel_1
-[csv]
-csv = ucayali_sentinel_granules_2018.csv
-[switches]
-submit_granules = True
-copy_processed_granules_to_bucket = True
-calculate_temporal_average = True
-dry_season_only = False
-[misc]
-max_threads = 4
